A couple of simple boards...
Here are a few of the simple boards I have used with playing with the OpenCM 9.04 boards.  I am mainly trying them to work with the Trossen Robotics Hexapod and the like who have mounting hole patterns which work with these boards.

This includes a few different versions of XBee adapters, that should be able to use a simple cable to plug into the modem/Uart connector of the 904C board.

I also have a couple fo Dynamixel MUX boards, which are similar to the Trossen Robotics MUX with power connectors as well as 6 DXL connectors.  In addition these have an XL320 connector, which should allow you to use simple servo extension cable from here to one of the XL320 connectors on the OpenCM boards, to both power the board as well as have the 904 board provide the DXL signal (Serial1).

There are two version of the MUX.  One with AX/MX servo connections and another with XL420 like connections, so you can connect up to XL420 like servos without using an AX to DXL servo connector wire...